requirements_saliency_ad.txt
absl-py==1.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Django==2.2
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.5
entrypoints==0.4
Faker==15.1.1
fastjsonschema==2.15.3
flatbuffers==23.3.3
fonttools==4.34.4
frozenlist==1.3.3
gast==0.2.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
greenlet==2.0.1
grpcio==1.47.0
gTTS==2.3.1
h5py==2.10.0
httplib2==0.22.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Keras==2.2.2
Keras-Applications==1.0.7
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.0.2
keras-vis @ git+https://github.com/raghakot/keras-vis.git@90ae5565951b5e6a90d706b8205c2c4dfc271505
kiwisolver==1.4.3
libclang==14.0.1
loguru==0.6.0
Markdown==3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
multidict==6.0.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.12
numpy==1.18.5
oauthlib==3.2.0
openai==0.27.4
opencv-python==4.6.0.66
opt-einsum==3.3.0
orjson==3.8.10
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
pinecone-client==2.2.1
playsound==1.3.0
playwright==1.32.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.1.0
seaborn==0.12.2
semver==2.13.0
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
streamlit==1.19.0
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.15.0
tensorflow-cpu==2.11.0
tensorflow-estimator==1.15.1
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2021.11.2
TikTokApi==5.2.2
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
torch==1.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
validators==0.20.0
vis==0.0.5
vocab==0.0.5
watchdog==2.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
win32-setctime==1.1.0
wrapt==1.12.1
yarl==1.8.2
zipp==3.8.1
